Recursion time reduced
Now the runs in my computer indicate that the info for edman calculation are the bottlenecks in the recursion, and a faster runtime. The init computing times and the picking the sequence are considerable (around 1/3). I think this can be further reduced so ifed calc becomes the bottleneck of the whole classification